+++ common/src/main/java/net/mehvahdjukaar/supplementaries/integration/SoulFiredCompat.java
@@ -3,13 +3,12 @@ package net.mehvahdjukaar.supplementaries.integration;
 import it.crystalnest.soul_fire_d.api.FireManager;
 import net.mehvahdjukaar.supplementaries.Supplementaries;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.RandomizableContainer;
 import net.minecraft.world.entity.Entity;
 
 public class SoulFiredCompat {
     public static final ResourceLocation LUMISENE_FIRE_TYPE = Supplementaries.res("lumisene");
 
-    public static void init() {
+    public static void setup() {
         FireManager.registerFire(
                 FireManager.fireBuilder(LUMISENE_FIRE_TYPE)
                         .setDamage(1)