+++ changelog.md
@@ -1,4 +1,3 @@
 CHECK 3.0.0 PATCH NOTES!!
 
-- fixed presents not saving their content
-- fixed forge render bounding box of some tile entities not being enlarged like it should
+- updated soul fired compat